:sparkles: 订单服务编码

+++ src/main/java/com/fenghuang/job/entity/CashWithdrawal.java
@@ -3,6 +3,7 @@ package com.fenghuang.job.entity;
 import io.swagger.annotations.ApiModel;
 import lombok.Data;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -14,7 +15,7 @@ public class CashWithdrawal implements Serializable{
 
     private Integer userId;
 
-    private Long cashWithdrawalAmount;
+    private BigDecimal cashWithdrawalAmount;
 
     private String cashWithdrawalAccount;
