+++ src/main/java/com/fenghuang/job/view/CashWithdrawalView.java
@@ -3,6 +3,7 @@ package com.fenghuang.job.view;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -19,7 +20,7 @@ public class CashWithdrawalView implements Serializable{
 
     private Integer userId;
 
-    private Long cashWithdrawalAmount;
+    private BigDecimal cashWithdrawalAmount;
 
     private String cashWithdrawalAccount;
